Initial cone tracer

+++ src/ShaderConstants.h
@@ -18,7 +18,7 @@ struct PerFrameUBO
     float padding2;
     glm::vec3 uCamUp;
     float padding3;
-    glm::uvec2 uResolution;
+    glm::vec2 uResolution;
     float uTime;
 };
